Added more realistic deceleration and acceleration
Deceleration is now inversely proportional to distance and directly
proportional to relative velocity. There is a maximum deceleration
which is a fraction of g. All balls are treated equal in that sense, i.e.
all vehicles have same braking characteristics which is probably not
right, for instance, bicycles decelerate less than a car before throwing
the rider over the handlebars.

Similarly acceleration is now a small fraction of g.